Split vcard plugin up into multiple files
